+++ plugins/module_utils/table.py
@@ -63,6 +63,14 @@ class TableClient:
 
         return records[0] if records else None
 
+    def get_record_by_sys_id(self, table, sys_id):
+        response = self.client.get(
+            _path(table, sys_id)
+        )
+        record = response.json["result"]
+
+        return record
+
     def create_record(self, table, payload, check_mode, query=None):
         if check_mode:
             # Approximate the result using the payload.